refactor: delete src/themes/common-variables.js
- add src/constants/colors.js
- add src/constants/typography.js
- modify files which import common-variables.js
