+++ src/components/about-us/section-02/departments-name-list.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../themes/common-variables'
+import colors from '../../../constants/colors'
 import { font } from '../constants/styles'
 import { gray } from './utils'
 import PropTypes from 'prop-types'
