+++ src/components/about-us/utils/navigation.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../themes/common-variables'
+import colors from '../../../constants/colors'
 import mq from './media-query'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
