+++ src/constants/colors.js
@@ -1,24 +1,4 @@
-export const typography = {
-  font: {
-    size: {
-      xLarger: '34px',
-      larger: '22px',
-      base: '18px',
-      medium: '16px',
-      small: '15px',
-      xSmall: '14px',
-      xSmaller: '13px',
-      tiny: '12px'
-    },
-    weight: {
-      extraLight: '100',
-      normal: '300',
-      bold: '700'
-    }
-  }
-}
-
-export const colors = {
+export default {
   black: '#000',
   white: '#fff',
   gray: {
@@ -44,8 +24,3 @@ export const colors = {
   secondaryColor: '#a67a44'
 }
 
-export default {
-  colors,
-  lineHeight,
-  typography
-}
