+++ src/constants/typography.js
@@ -1,0 +1,19 @@
+export default {
+  font: {
+    size: {
+      xLarger: '34px',
+      larger: '22px',
+      base: '18px',
+      medium: '16px',
+      small: '15px',
+      xSmall: '14px',
+      xSmaller: '13px',
+      tiny: '12px'
+    },
+    weight: {
+      extraLight: '100',
+      normal: '300',
+      bold: '700'
+    }
+  }
+}
